requirements.txt
altair==5.0.1
extra-streamlit-components==0.1.56
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.3
numpy-financial==1.0.0
openpyxl==3.1.0
pandas==1.3.5
Pillow==9.5.0
platformdirs==3.5.1
seaborn==0.12.2
streamlit==1.27.0
streamlit-extras==0.3.2
streamlit-option-menu==0.3.6
plotly

